feat: 권한 검증이 필요한 곳에서 권한 검증을 수행한다 (#112)
* [#111] feat: About 권한 검증 추가

* [#111] feat: Category 권한 검증 추가

* [#111] feat: Blog, Post 권한 검증 추가

* [#111] feat: Comment 권한 검증 추가

* [#111] refactor: polishing

+++ src/main/java/com/mallang/category/application/CategoryService.java
@@ -27,21 +27,23 @@ public class CategoryService {
     public Long create(CreateCategoryCommand command) {
         Member member = memberRepository.getById(command.memberId());
         Blog blog = blogRepository.getByNameAndOwner(command.blogName(), command.memberId());
-        Category parentCategory = categoryRepository
-                .getParentByIdAndOwner(command.parentCategoryId(), command.memberId());
+        Category parentCategory = categoryRepository.getParentById(command.parentCategoryId());
         Category category = Category.create(command.name(), member, blog, parentCategory, categoryValidator);
         return categoryRepository.save(category).getId();
     }
 
     public void update(UpdateCategoryCommand command) {
-        Category category = categoryRepository.getByIdAndOwner(command.categoryId(), command.memberId());
-        Category parentCategory = categoryRepository
-                .getParentByIdAndOwner(command.parentCategoryId(), command.memberId());
+        Member member = memberRepository.getById(command.memberId());
+        Category category = categoryRepository.getById(command.categoryId());
+        category.validateOwner(member);
+        Category parentCategory = categoryRepository.getParentById(command.parentCategoryId());
         category.update(command.name(), parentCategory, categoryValidator);
     }
 
     public void delete(DeleteCategoryCommand command) {
-        Category category = categoryRepository.getByIdAndOwner(command.categoryId(), command.memberId());
+        Member member = memberRepository.getById(command.memberId());
+        Category category = categoryRepository.getById(command.categoryId());
+        category.validateOwner(member);
         category.delete();
         categoryRepository.delete(category);
     }
